Style form, allow for No! ingredients!
Moved styling into its own file. Created a generic .table class for
some normalisation styles, along with making some styles slightly
more consistent (the inputs/buttons). Decided to leave it all in the
one file for now. Decided to get make a table, get rid of it if no
fields exist, with a simple "Ingredients" title. I played around with
a few different ways of presenting this, and this one made the most
sense.

Allowed for no ingredients to be passed through. React-hook-form gets
rid of the whole key if no ingredients exist. Decided not to make the
field optional on the server, it feels more appropriate to add on the
client. Re-appropriate the cleanData function as it is the simplest
way of doing what we want! Can slowly clean up the test data and turn
into a nice little helper function.